Move sql render context provider test support class to a top-level class
