+++ src/test/java/bio/terra/service/snapshotbuilder/query/OrderByVariableTest.java
@@ -20,21 +20,21 @@ class OrderByVariableTest {
   }
 
   @ParameterizedTest
-  @ArgumentsSource(SqlRenderContextTest.Contexts.class)
+  @ArgumentsSource(SqlRenderContextProvider.class)
   void renderSQLAsc(SqlRenderContext context) {
     var orderByVariable = new OrderByVariable(createVariable());
     assertThat(orderByVariable.renderSQL(false, context), is("t.column ASC"));
   }
 
   @ParameterizedTest
-  @ArgumentsSource(SqlRenderContextTest.Contexts.class)
+  @ArgumentsSource(SqlRenderContextProvider.class)
   void renderSQLDesc(SqlRenderContext context) {
     var orderByVariable = new OrderByVariable(createVariable(), OrderByDirection.DESCENDING);
     assertThat(orderByVariable.renderSQL(false, context), is("t.column DESC"));
   }
 
   @ParameterizedTest
-  @ArgumentsSource(SqlRenderContextTest.Contexts.class)
+  @ArgumentsSource(SqlRenderContextProvider.class)
   void renderSQLRandom(SqlRenderContext context) {
     var orderByVariable = OrderByVariable.random();
     assertThat(orderByVariable.renderSQL(true, context), is("RAND()"));
