+++ src/test/java/bio/terra/service/snapshotbuilder/query/TablePointerTest.java
@@ -13,7 +13,7 @@ import org.junit.jupiter.params.provider.ArgumentsSource;
 class TablePointerTest {
 
   @ParameterizedTest
-  @ArgumentsSource(SqlRenderContextTest.Contexts.class)
+  @ArgumentsSource(SqlRenderContextProvider.class)
   void renderSQL(SqlRenderContext context) {
     var tablePointer =
         new TablePointer("table", (primaryTable, tables) -> (cloudPlatform) -> "filter");
