+++ src/test/java/bio/terra/service/snapshotbuilder/query/UnionQueryTest.java
@@ -15,7 +15,7 @@ import org.junit.jupiter.params.provider.ArgumentsSource;
 class UnionQueryTest {
 
   @ParameterizedTest
-  @ArgumentsSource(SqlRenderContextTest.Contexts.class)
+  @ArgumentsSource(SqlRenderContextProvider.class)
   void renderSQL(SqlRenderContext context) {
     var query = QueryTest.createQuery();
     var unionQuery = new UnionQuery(List.of(query, query));
