+++ src/test/java/bio/terra/service/snapshotbuilder/query/filtervariable/BooleanAndOrFilterVariableTest.java
@@ -6,7 +6,7 @@ import static org.hamcrest.Matchers.is;
 import bio.terra.common.category.Unit;
 import bio.terra.service.snapshotbuilder.query.Literal;
 import bio.terra.service.snapshotbuilder.query.SqlRenderContext;
-import bio.terra.service.snapshotbuilder.query.SqlRenderContextTest;
+import bio.terra.service.snapshotbuilder.query.SqlRenderContextProvider;
 import bio.terra.service.snapshotbuilder.query.TablePointer;
 import bio.terra.service.snapshotbuilder.query.TableVariable;
 import org.junit.jupiter.api.Tag;
@@ -34,13 +34,13 @@ class BooleanAndOrFilterVariableTest {
   }
 
   @ParameterizedTest
-  @ArgumentsSource(SqlRenderContextTest.Contexts.class)
+  @ArgumentsSource(SqlRenderContextProvider.class)
   void renderSQL(SqlRenderContext context) {
     assertThat(variable.renderSQL(context), is("(t.field1 = 'value1' AND t1.field2 = 'value2')"));
   }
 
   @ParameterizedTest
-  @ArgumentsSource(SqlRenderContextTest.Contexts.class)
+  @ArgumentsSource(SqlRenderContextProvider.class)
   void renderSQLWorksWithNoSubQueries(SqlRenderContext renderContext) {
     assertThat(BooleanAndOrFilterVariable.and().renderSQL(renderContext), is("1=1"));
   }
