+++ src/test/java/bio/terra/service/snapshotbuilder/query/filtervariable/FunctionFilterVariableTest.java
@@ -8,7 +8,7 @@ import bio.terra.service.snapshotbuilder.query.FieldPointer;
 import bio.terra.service.snapshotbuilder.query.FieldVariable;
 import bio.terra.service.snapshotbuilder.query.Literal;
 import bio.terra.service.snapshotbuilder.query.SqlRenderContext;
-import bio.terra.service.snapshotbuilder.query.SqlRenderContextTest;
+import bio.terra.service.snapshotbuilder.query.SqlRenderContextProvider;
 import bio.terra.service.snapshotbuilder.query.TablePointer;
 import bio.terra.service.snapshotbuilder.query.TableVariable;
 import org.junit.jupiter.api.Tag;
@@ -19,7 +19,7 @@ import org.junit.jupiter.params.provider.ArgumentsSource;
 class FunctionFilterVariableTest {
 
   @ParameterizedTest
-  @ArgumentsSource(SqlRenderContextTest.Contexts.class)
+  @ArgumentsSource(SqlRenderContextProvider.class)
   void renderSQL(SqlRenderContext context) {
     TableVariable table = TableVariable.forPrimary(TablePointer.fromTableName("table"));
     var filterVariable =
